+++ src/arch/Linux-x86_64/syscall.c
@@ -1,22 +1,24 @@
-unsigned long _asm_syscall0(int number)
+#include <sys/asm/syscall.h>
+
+uintptr_t _asm_syscall0(int number)
 {
-    unsigned long ret;
+    uintptr_t ret;
 
     asm volatile (
             "movq %1, %%rax\n\t"
             "syscall\n\t"
             "movq %%rax, %0\n\t"
             :"=r"(ret)
-            :"r"((unsigned long)number)
+            :"r"((uintptr_t)number)
             :"%rax","%rcx","%r11"
        );
 
     return ret;
 }
 
-unsigned long _asm_syscall1(int number, unsigned long arg0)
+uintptr_t _asm_syscall1(int number, uintptr_t arg0)
 {
-    unsigned long ret;
+    uintptr_t ret;
 
     asm volatile (
             "movq %1, %%rax\n\t"
@@ -24,17 +26,17 @@ unsigned long _asm_syscall1(int number, unsigned long arg0)
             "syscall\n\t"
             "movq %%rax, %0\n\t"
             :"=r"(ret)
-            :"r"((unsigned long)number),"r"(arg0)
+            :"r"((uintptr_t)number),"r"(arg0)
             :"%rax","%rdi","%rcx","%r11"
        );
 
     return ret;
 }
 
-unsigned long _asm_syscall2(int number, unsigned long arg0,
-        unsigned long arg1)
+uintptr_t _asm_syscall2(int number, uintptr_t arg0,
+        uintptr_t arg1)
 {
-    unsigned long ret;
+    uintptr_t ret;
 
     asm volatile (
             "movq %1, %%rax\n\t"
@@ -43,17 +45,17 @@ unsigned long _asm_syscall2(int number, unsigned long arg0,
             "syscall\n\t"
             "movq %%rax, %0\n\t"
             :"=r"(ret)
-            :"r"((unsigned long)number),"r"(arg0),"r"(arg1)
+            :"r"((uintptr_t)number),"r"(arg0),"r"(arg1)
             :"%rax","%rdi","%rsi","%rcx","%r11"
        );
 
     return ret;
 }
 
-unsigned long _asm_syscall3(int number, unsigned long arg0,
-        unsigned long arg1, unsigned long arg2)
+uintptr_t _asm_syscall3(int number, uintptr_t arg0,
+        uintptr_t arg1, uintptr_t arg2)
 {
-    unsigned long ret;
+    uintptr_t ret;
 
     asm volatile (
             "movq %1, %%rax\n\t"
@@ -63,17 +65,17 @@ unsigned long _asm_syscall3(int number, unsigned long arg0,
             "syscall\n\t"
             "movq %%rax, %0\n\t"
             :"=r"(ret)
-            :"r"((unsigned long)number),"r"(arg0),"r"(arg1),"r"(arg2)
+            :"r"((uintptr_t)number),"r"(arg0),"r"(arg1),"r"(arg2)
             :"%rax","%rdi","%rsi","%rdx","%rcx","%r11"
        );
 
     return ret;
 }
 
-unsigned long _asm_syscall4(int number, unsigned long arg0,
-        unsigned long arg1, unsigned long arg2, unsigned long arg3)
+uintptr_t _asm_syscall4(int number, uintptr_t arg0,
+        uintptr_t arg1, uintptr_t arg2, uintptr_t arg3)
 {
-    unsigned long ret;
+    uintptr_t ret;
 
     asm volatile (
             "movq %1, %%rax\n\t"
@@ -84,7 +86,7 @@ unsigned long _asm_syscall4(int number, unsigned long arg0,
             "syscall\n\t"
             "movq %%rax, %0\n\t"
             :"=r"(ret)
-            :"r"((unsigned long)number),"r"(arg0),"r"(arg1),"r"(arg2),
+            :"r"((uintptr_t)number),"r"(arg0),"r"(arg1),"r"(arg2),
             "r"(arg3)
             :"%rax","%rdi","%rsi","%rdx","%r10","%rcx","%r11"
        );
@@ -92,11 +94,11 @@ unsigned long _asm_syscall4(int number, unsigned long arg0,
     return ret;
 }
 
-unsigned long _asm_syscall5(int number, unsigned long arg0,
-        unsigned long arg1, unsigned long arg2, unsigned long arg3,
-        unsigned long arg4)
+uintptr_t _asm_syscall5(int number, uintptr_t arg0,
+        uintptr_t arg1, uintptr_t arg2, uintptr_t arg3,
+        uintptr_t arg4)
 {
-    unsigned long ret;
+    uintptr_t ret;
 
     asm volatile (
             "movq %1, %%rax\n\t"
@@ -108,7 +110,7 @@ unsigned long _asm_syscall5(int number, unsigned long arg0,
             "syscall\n\t"
             "movq %%rax, %0\n\t"
             :"=r"(ret)
-            :"r"((unsigned long)number),"r"(arg0),"r"(arg1),"r"(arg2),
+            :"r"((uintptr_t)number),"r"(arg0),"r"(arg1),"r"(arg2),
             "r"(arg3),"r"(arg4)
             :"%rax","%rdi","%rsi","%rdx","%r10","%r8","%rcx","%r11"
        );
@@ -116,11 +118,11 @@ unsigned long _asm_syscall5(int number, unsigned long arg0,
     return ret;
 }
 
-unsigned long _asm_syscall6(int number, unsigned long arg0,
-        unsigned long arg1, unsigned long arg2, unsigned long arg3,
-        unsigned long arg4, unsigned long arg5)
+uintptr_t _asm_syscall6(int number, uintptr_t arg0,
+        uintptr_t arg1, uintptr_t arg2, uintptr_t arg3,
+        uintptr_t arg4, uintptr_t arg5)
 {
-    unsigned long ret;
+    uintptr_t ret;
 
     asm volatile (
             "pushq %%r11\n\t"
@@ -135,7 +137,7 @@ unsigned long _asm_syscall6(int number, unsigned long arg0,
             "popq %%r11\n\t"
             "movq %%rax, %0\n\t"
             :"=r"(ret)
-            :"r"((unsigned long)number),"r"(arg0),"r"(arg1),"r"(arg2),
+            :"r"((uintptr_t)number),"r"(arg0),"r"(arg1),"r"(arg2),
             "r"(arg3),"r"(arg4),"r"(arg5)
             :"%rax","%rdi","%rsi","%rdx","%r10","%r8","%r9","%rcx"
        );
